bugs fixed

+++ src/controllers/video.controller.js
@@ -4,36 +4,81 @@ import { User } from "../models/user.model.js";
 import { ApiError } from "../utils/ApiError.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
-import { uploadOnCloudinary } from "../utils/cloudinary.js";
+import { uploadOnCloudinary,deleteImageFromCloudinary } from "../utils/cloudinary.js";
 
 const getAllVideos = asyncHandler(async (req, res) => {
   const { page = 1, limit = 10, query, sortBy, sortType, userId } = req.query;
   //TODO: get all videos based on query, sort, pagination
 
 try {
-      const aggregate = Video.aggregate([
-        {
-          $match: {
-            owner: new mongoose.Types.ObjectId(userId),
+  const pipeline = [];
+
+  if (!userId) {
+    throw new ApiError(400, "User Id not provided");
+  }
+
+  if (!isValidObjectId(userId)) {
+    throw new ApiError(400, "Invalid User Id");
+  }
+
+  if (userId) {
+    pipeline.push({
+      $match: {
+        owner: new mongoose.Types.ObjectId(userId),
+      },
+    });
+  }
+
+  if (query) {
+    pipeline.push({
+      $match: {
+        title: {
+          $regex: query,
+        },
+      },
+    });
+  }
+
+  pipeline.push({ $match: { isPublished: true } });
+
+  pipeline.push(
+    {
+      $lookup: {
+        from: "users",
+        localField: "owner",
+        foreignField: "_id",
+        as: "owner",
+        pipeline: [
+          {
+            $project: {
+              username: 1,
+              fullname: 1,
+              avatar: 1,
+            },
           },
-        },
-      ]);
+        ],
+      },
+    },
+    {
+      $unwind: "$owner",
+    }
+  );
 
-      const options = {
-        page,
-        limit,
-        sort: {
-          [sortBy]: sortType,
-        },
-      };
+  const aggregate = Video.aggregate(pipeline);
 
-      const videos=await Video.aggregatePaginate(aggregate, options);
+  const options = {
+    page,
+    limit,
+    sort: {
+      [sortBy]: sortType,
+    },
+  };
 
-      return res
-        .status(200)
-        .json(
-          new ApiResponse(200, videos.docs, "Videos fetched successfully")
-        );
+  const videos = await Video.aggregatePaginate(aggregate, options);
+
+  return res
+    .status(200)
+    .json(new ApiResponse(200, videos.docs, "Videos fetched successfully"));
 } catch (error) {
      return res
        .status(500)
@@ -128,11 +173,17 @@ const updateVideo = asyncHandler(async (req, res) => {
     throw new ApiError(400, "thumbnail is missing");
   }
 
+  const oldVideo = await Video.findById(videoId);
+
+  const oldThumbnail = oldVideo.thumbnail;
+
   const newThumbnail = await uploadOnCloudinary(thumbnailLocalPath);
 
   if (!newThumbnail) {
     throw new ApiError(400, "thumbnail is missing");
   }
+
+  await deleteImageFromCloudinary(oldThumbnail);
 
   const updatedVideo = await Video.findByIdAndUpdate(
     videoId,
@@ -145,6 +196,10 @@ const updateVideo = asyncHandler(async (req, res) => {
     },
     { new: true }
   );
+
+  if (!updatedVideo) {
+    throw new ApiError(500, "Video not found after updating the details");
+  }
 
   return res
     .status(200)
@@ -160,7 +215,15 @@ const deleteVideo = asyncHandler(async (req, res) => {
     throw new ApiError(400, "Invalid video id");
   }
 
+  const video = await Video.findById(videoId);
+  if (!video) {
+    throw new ApiError(400, "No video with given id exists");
+  }
+
   await Video.findByIdAndDelete(videoId);
+
+  await deleteImageFromCloudinary(video.videoFile);
+  await deleteImageFromCloudinary(video.thumbnail);
 
   return res
     .status(200)
@@ -175,6 +238,9 @@ const togglePublishStatus = asyncHandler(async (req, res) => {
   }
 
   const video = await Video.findById(videoId);
+  if (!video) {
+    throw new ApiError(404, "Video not found");
+  }
 
   video.isPublished = !video.isPublished;
   await video.save({ validateBeforeSave: false });
